Урок 2. Каждая задача вынесена в отдельный файл.
